+++ src/routes/(v2)/Navigation.svelte
@@ -16,6 +16,7 @@
 		<li><a href="/v2/letters/asdf">Letter route</a></li>
 		<li><a href="/v2/custom-tainted-dialog">Custom tainted dialog</a></li>
 		<li><a href="/v2/issue-309-unions">Schema union</a></li>
+		<li><a href="/v2/on-change-event">onChange event</a></li>
 	</ul>
 </nav>