+++ src/routes/(v2)/v2/on-change-event/schema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	name: z.string().min(1),
+	email: z.string().email(),
+	agree: z.literal(true).default(false as true)
+});